+++ src/appInterface.h
@@ -28,12 +28,11 @@ class IApp {
 
         virtual void addOnce(ah::scdCb c, uint32_t timeout, const char *name) = 0;
 
-        #if !defined(ETHERNET)
         virtual bool getAvailNetworks(JsonObject obj) = 0;
         virtual void setupStation(void) = 0;
         virtual bool getWasInCh12to14(void) const = 0;
-        #endif /* defined(ETHERNET) */
         virtual String getIp(void) = 0;
+        virtual String getMac(void) = 0;
         virtual bool isApActive(void) = 0;
 
         virtual uint32_t getUptime() = 0;
@@ -58,6 +57,10 @@ class IApp {
 
         virtual uint32_t getMqttRxCnt() = 0;
         virtual uint32_t getMqttTxCnt() = 0;
+
+        #if defined(ETHERNET)
+        virtual bool isWiredConnection() = 0;
+        #endif
 
         virtual void lock(bool fromWeb) = 0;
         virtual char *unlock(const char *clientIp, bool loginFromWeb) = 0;