+++ src/network/AhoyEthernet.h
@@ -12,11 +12,20 @@
 #include <ETH.h>
 #include "AhoyEthernetSpi.h"
 #include "AhoyNetwork.h"
+#include "AhoyWifiEsp32.h"
 
-class AhoyEthernet : public AhoyNetwork {
+class AhoyEthernet : public AhoyWifi {
+    private:
+        enum class Mode {
+            WIRED,
+            WIRELESS
+        };
+
     public:
-        void begin() override {
+        virtual void begin() override {
+            mMode = Mode::WIRELESS;
             mAp.enable();
+            AhoyWifi::begin();
 
             if(!mConfig->sys.eth.enabled)
                 return;
@@ -25,24 +34,66 @@ class AhoyEthernet : public AhoyNetwork {
             ETH.setHostname(mConfig->sys.deviceName);
         }
 
-        void OnEvent(WiFiEvent_t event) override {
+        virtual String getIp(void) override {
+            if(Mode::WIRELESS == mMode)
+                return AhoyWifi::getIp();
+            else
+                return ETH.localIP().toString();
+        }
+
+        virtual String getMac(void) override {
+            if(Mode::WIRELESS == mMode)
+                return AhoyWifi::getMac();
+            else
+                return mEthSpi.macAddress();
+        }
+
+        virtual bool isWiredConnection() override {
+            return (Mode::WIRED == mMode);
+        }
+
+    private:
+        virtual void OnEvent(WiFiEvent_t event) override {
             switch(event) {
                 case ARDUINO_EVENT_ETH_CONNECTED:
+                    mMode = Mode::WIRED; // needed for static IP
+                    [[fallthrough]];
+                case SYSTEM_EVENT_STA_CONNECTED:
+                    mWifiConnecting = false;
                     if(NetworkState::CONNECTED != mStatus) {
+                        if(ARDUINO_EVENT_ETH_CONNECTED == event)
+                            WiFi.disconnect();
+
                         mStatus = NetworkState::CONNECTED;
                         DPRINTLN(DBG_INFO, F("Network connected"));
                         setStaticIp();
                     }
                     break;
 
+                case SYSTEM_EVENT_STA_GOT_IP:
+                    mStatus = NetworkState::GOT_IP;
+                    if(mAp.isEnabled())
+                        mAp.disable();
+
+                    mMode = Mode::WIRELESS;
+                    if(!mConnected) {
+                        mConnected = true;
+                        ah::welcome(WiFi.localIP().toString(), F("Station WiFi"));
+                        MDNS.begin(mConfig->sys.deviceName);
+                        mOnNetworkCB(true);
+                    }
+                    break;
+
                 case ARDUINO_EVENT_ETH_GOT_IP:
                     mStatus = NetworkState::GOT_IP;
+                    mMode = Mode::WIRED;
                     if(!mConnected) {
                         mAp.disable();
                         mConnected = true;
-                        ah::welcome(ETH.localIP().toString(), F("Station"));
+                        ah::welcome(ETH.localIP().toString(), F("Station Ethernet"));
                         MDNS.begin(mConfig->sys.deviceName);
                         mOnNetworkCB(true);
+                        WiFi.disconnect();
                     }
                     break;
 
@@ -51,9 +102,25 @@ class AhoyEthernet : public AhoyNetwork {
                 case ARDUINO_EVENT_ETH_DISCONNECTED:
                     mStatus = NetworkState::DISCONNECTED;
                     if(mConnected) {
+                        mMode = Mode::WIRELESS;
                         mConnected = false;
                         mOnNetworkCB(false);
-                        mAp.enable();
+                        MDNS.end();
+                        AhoyWifi::begin();
+                    }
+                    break;
+
+                case ARDUINO_EVENT_WIFI_STA_LOST_IP:
+                    [[fallthrough]];
+                case ARDUINO_EVENT_WIFI_STA_STOP:
+                    [[fallthrough]];
+                case SYSTEM_EVENT_STA_DISCONNECTED:
+                    mStatus = NetworkState::DISCONNECTED;
+                    if(mConnected && (Mode::WIRELESS == mMode)) {
+                        mConnected = false;
+                        mOnNetworkCB(false);
+                        MDNS.end();
+                        AhoyWifi::begin();
                     }
                     break;
 
@@ -62,24 +129,19 @@ class AhoyEthernet : public AhoyNetwork {
             }
         }
 
-        void tickNetworkLoop() override {
-            if(mAp.isEnabled())
-                mAp.tickLoop();
-        }
-
-        String getIp(void) override {
-            return ETH.localIP().toString();
-        }
-
-    private:
         void setStaticIp() override {
             setupIp([this](IPAddress ip, IPAddress gateway, IPAddress mask, IPAddress dns1, IPAddress dns2) -> bool {
-                return ETH.config(ip, gateway, mask, dns1, dns2);
+                if(Mode::WIRELESS == mMode)
+                    return WiFi.config(ip, gateway, mask, dns1, dns2);
+                else
+                    return ETH.config(ip, gateway, mask, dns1, dns2);
             });
         }
 
     private:
         AhoyEthernetSpi mEthSpi;
+        Mode mMode;
+
 };
 
 #endif /*ETHERNET*/