+++ src/network/AhoyNetwork.h
@@ -79,8 +79,13 @@ class AhoyNetwork {
         virtual void begin() = 0;
         virtual void tickNetworkLoop() = 0;
         virtual String getIp(void) = 0;
+        virtual String getMac(void) = 0;
 
         virtual bool getWasInCh12to14() {
+            return false;
+        }
+
+        virtual bool isWiredConnection() {
             return false;
         }
 
@@ -88,7 +93,6 @@ class AhoyNetwork {
             return mAp.isEnabled();
         }
 
-        #if !defined(ETHERNET)
         bool getAvailNetworks(JsonObject obj, IApp *app) {
             if(!mScanActive) {
                 app->addOnce([this]() {scan();}, 1, "scan");
@@ -122,7 +126,6 @@ class AhoyNetwork {
             }
             WiFi.scanNetworks(true, true);
         }
-        #endif
 
     protected:
         virtual void setStaticIp() = 0;
@@ -173,7 +176,6 @@ class AhoyNetwork {
             }
         }
 
-        #if !defined(ETHERNET)
         void sortRSSI(int *sort, int n) {
             for (int i = 0; i < n; i++)
                 sort[i] = i;
@@ -182,7 +184,6 @@ class AhoyNetwork {
                     if (WiFi.RSSI(sort[j]) > WiFi.RSSI(sort[i]))
                         std::swap(sort[i], sort[j]);
         }
-        #endif
 
     private:
         void sendNTPpacket(void) {
@@ -234,9 +235,7 @@ class AhoyNetwork {
         uint32_t *mUtcTimestamp = nullptr;
         bool mConnected = false;
         bool mScanActive = false;
-        #if !defined(ETHERNET)
         bool mWifiConnecting = false;
-        #endif
 
         OnNetworkCB mOnNetworkCB;
         OnTimeCB mOnTimeCB;